+++ docs/CHANGELOG.md
@@ -29,3 +29,4 @@
 - [Simeon Sodunke] Updated Player model, view and controller
 -[Samuel Yohannes] Added Games Scores Class
 -[Melkamu Gonfa]  added and updated model, view and controller for the Game class
+- [Noah Manning] Updated view, model, and controller for the Sport class